Make span size unsigned (as in c++20 now)

+++ docs/tests/span.cpp
@@ -71,5 +71,5 @@ TEST(span) {
 
 	count = 0;
 	foo({namesVector.data(), 4}, count);
-	foo({namesVector.data(), (int) namesVector.size()}, count);
+	foo({namesVector.data(), namesVector.size()}, count);
 }
